5th commit
First of all I started naming naming the commits a little differently to make the easier to read. I have made the HTML a little more easy to read and updated the blank HTML. And as always I made some little changes to the code. In the next commit I'd like to get rid of unnecessary CSS code.